Пересчёт TX-кадра и переход на 16-битные идентификаторы

+++ libs/packetizer/packet_splitter.h
@@ -22,7 +22,7 @@ public:
   void setCustomSize(size_t custom);
   // Разделить данные и занести части в буфер, возвращает ID первого пакета
   // with_id=true добавляет префикс и журналирование
-  uint32_t splitAndEnqueue(MessageBuffer& buf, const uint8_t* data, size_t len, bool with_id = true) const;
+  uint16_t splitAndEnqueue(MessageBuffer& buf, const uint8_t* data, size_t len, bool with_id = true) const;
   // Возвращает размер полезной нагрузки для текущего режима
   size_t payloadSize() const;
 private:
