+++ message_buffer.h
@@ -13,7 +13,7 @@ public:
   static constexpr size_t DEFAULT_SLOT_SIZE = 256; // размер слота по умолчанию
   explicit MessageBuffer(size_t capacity, size_t slot_size = DEFAULT_SLOT_SIZE);
   // Добавляет сообщение в буфер и возвращает его идентификатор
-  uint32_t enqueue(const uint8_t* data, size_t len);
+  uint16_t enqueue(const uint8_t* data, size_t len);
   // Возвращает количество свободных слотов в очереди
   size_t freeSlots() const;
   // Удаляет последнее сообщение (для отката операций)
@@ -21,19 +21,19 @@ public:
   // Проверяет, есть ли сообщения в очереди
   bool hasPending() const;
   // Извлекает первое сообщение из очереди, возвращает ID и данные
-  bool pop(uint32_t& id, std::vector<uint8_t>& out);
+  bool pop(uint16_t& id, std::vector<uint8_t>& out);
   // Позволяет заглянуть в начало очереди без извлечения
-  const std::vector<uint8_t>* peek(uint32_t& id) const;
+  const std::vector<uint8_t>* peek(uint16_t& id) const;
   // Возвращает вместимость слота данных
   size_t slotSize() const;
 private:
   struct Slot {
-    uint32_t id = 0;                  // сохранённый идентификатор сообщения
+    uint16_t id = 0;                  // сохранённый идентификатор сообщения
     std::vector<uint8_t> data;        // заранее выделенный буфер данных
     bool used = false;                // флаг заполнения слота
   };
 
-  uint32_t next_id_ = 1;             // следующий идентификатор
+  uint16_t next_id_ = 1;             // следующий идентификатор
   size_t capacity_;                  // максимальное количество сообщений
   size_t slot_size_;                 // максимальная длина данных одного сообщения
   std::vector<Slot> slots_;          // предвыделенные слоты хранения
